More somes fixes
Renamed "CALLBACK_ONADDFOCUS" for "CALLBACK_SET_INTERACTION"
And "CALLBACK_ONRELEASEFOCUS" for "CALLBACK_REMOVE_INTERACTION"

Removed focuses and idleTime, because npcs no longer need to check onThink if that is the focus, as there is no time limit to talk to them (unless it is kicked by idle)

Added the self:updateInteraction(npc, player) function in onCreatureSay so that every time you talk to npc, it clears the eventDelayedSay list, thus preventing npc from continuing to talk even after the player gives bye or tries to say something else.

Fixed the "NpcHandler:onMove" callback and added in "NpcHandler:say" the callback to send the message with other text, thus allowing the onWalkAway text pattern to be modified to be sent in regular chat instead of npc chat.